Move router/api into infrastructure
