+++ web_client/app/app-component.tsx
@@ -7,7 +7,7 @@ import { routes } from 'routes';
 import { Voter } from 'models';
 
 // Components
-import { Router } from 'router';
+import { Router } from 'infrastructure/router';
 import { NotFound } from 'pages/not-found/not-found';
 import * as DocumentTitle from 'react-document-title';
