+++ web_client/app/infrastructure/fullstory/fullstory-snippet.js
