+++ web_client/app/main.tsx
@@ -1,4 +1,4 @@
-require('fullstory/fullstory');
+require('infrastructure/fullstory/fullstory');
 import { AppContainer } from 'react-hot-loader';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
