+++ web_client/app/pages/login/login.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { observer } from 'mobx-react';
-import { go } from 'router';
+import { go } from 'infrastructure/router';
 import { Voter } from 'models';
 
 import { MainContentWrapper } from 'components/main-content-wrapper/main-content-wrapper';
@@ -10,7 +10,7 @@ import { Link } from 'components/link/link';
 
 import { styles, vars, css, centeredBox } from 'styles/css';
 
-import API from 'api/api';
+import API from 'infrastructure/api/api';
 
 export type LoginProps = {
