+++ web_client/app/pages/registration-check/registration-check.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { observer } from 'mobx-react';
 import { Voter } from 'models';
-import { go } from 'router';
+import { go } from 'infrastructure/router';
 
 import { Link } from 'components/link/link';
 import { MainContentWrapper } from 'components/main-content-wrapper/main-content-wrapper';
